# openapi.json
{"openapi":"3.1.0","info":{"title":"FlipFinder","version":"0.1.0"},"paths":{"/scrape/facebook":{"post":{"tags":["facebook"],"summary":"Run a one-off FB Marketplace scrape","operationId":"run_scrape_scrape_facebook_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScrapeRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Run Scrape Scrape Facebook Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/listing/{listing_id}/refresh_comps":{"post":{"summary":"Refresh Comps","operationId":"refresh_comps_listing__listing_id__refresh_comps_post","parameters":[{"name":"listing_id","in":"path","required":true,"schema":{"type":"integer","title":"Listing Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ScrapeRequest":{"properties":{"keywords":{"items":{"type":"string"},"type":"array","title":"Keywords"},"region":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Region","default":"Toronto, Ontario"},"radius_km":{"type":"integer","title":"Radius Km","default":120},"max_pages":{"type":"integer","maximum":10.0,"minimum":1.0,"title":"Max Pages","default":2},"auto_comps":{"type":"boolean","title":"Auto Comps","default":true},"email_on_deal":{"type":"boolean","title":"Email On Deal","default":true},"min_profit":{"type":"number","title":"Min Profit","default":75.0},"min_roi":{"type":"number","title":"Min Roi","default":0.25}},"type":"object","title":"ScrapeRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}